pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "umbra-gate"
version = "0.2.0"
description = "HTTP policy gate for autonomous agents -- powered by CI-1T"
requires-python = ">=3.10"
license = {text = "Elastic-2.0"}
authors = [{name = "Collapse Index"}]
readme = "README.md"
keywords = ["umbra", "ci1t", "agents", "monitoring", "safety", "guardrails"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "starlette>=0.36.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
slack = ["httpx"]
email = []
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx",
]

[project.scripts]
umbra = "umbra.__main__:main"

[project.urls]
Homepage = "https://collapseindex.org"
Repository = "https://github.com/collapseindex/umbra"
Documentation = "https://github.com/collapseindex/umbra#readme"
"API Keys" = "https://collapseindex.org/dashboard"

[tool.setuptools.packages.find]
include = ["umbra*"]
